+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * `/sign cut` only put empty lines into the clipboard (#17)
 * Line selection validation for `/sign cut` and `/sign copy`
+* Line selection validation accepts empty delimiters at the end (e.g. `/sign set 1,3,,,,`) but should not
 
 ## v1.12.2 (2020-08-10)